cdi-specs: make cdi spec public and refactor it
(1) As cdi spec will be involved in some other files,
we need make it and its related fields public.

(2) Change annotations: HashMap to BTreeMap, as the
object of Spec will be involved in HashSet with cmp
hash value in Cache. But HashMap has no impl for trait
Hash.

(3) Others,such as Mount.mount_type to r#type, and the
type of DeviceNode.file_mode as Option<mode_t>

Signed-off-by: Alex Lyn <alex.lyn@antgroup.com>